First pass at "language" attribute
Language attribute allows the user to set a language on the current
object and associate like-objects as translations. Currently, the
attribute is only implemented and tested on File objects, since the
package currently uses the MultilingualPageRelations table for the
pages, and I don't see a real practical use for the User object.
Ideally, C5 has future plans to allow easy addition of new
AttributeKeyCategories, and then the attribute can be used to define
locality and translations on arbitrary objects.